+++ src/components/pages/departments.js
@@ -18,6 +18,8 @@ import {
   ColorButton,
 } from "@material-ui/core";
 import { green } from "@material-ui/core/colors";
+import { ConfirmDelete } from "../alerts";
+
 const Departments = () => {
   const user = JSON.parse(localStorage.getItem("userData"));
 
@@ -67,20 +69,11 @@ const Departments = () => {
   };
 
   const deleteDepartmentSelected = async (departmentId) => {
-    const createResponse = await deleteDepartment(departmentId);
-
-    if (createResponse.code === 200) {
-      alert(createResponse.message);
-      refreshPage();
-    }
-
-    if (createResponse.code === 400) {
-      alert(createResponse.message);
-    }
-
-    if (createResponse.status === 401) {
-      alert(createResponse.error);
-    }
+    ConfirmDelete(
+      "¿Estás seguro de eliminar este departamento?",
+      "No podrás deshacer esta acción",
+      deleteDepartment(departmentId)
+    );
   };
 
   if (loading) {
@@ -181,52 +174,3 @@ const Departments = () => {
 };
 
 export default Departments;
-
-/*
-<div>
-      <button
-        onClick={() =>
-          history.push("/municipalidades/departamentos/crear", {
-            from: "municipalidades/departamentos",
-          })
-        }
-      >
-        Crear
-      </button>
-
-      <button
-        onClick={() => {
-          history.push("/municipalidades");
-        }}
-      >
-        Volver
-      </button>
-
-      <h2>Depts</h2>
-
-      <ul>
-        {departments.map((department) => (
-          <ul key={department.id}>
-            <h2>{department.name}</h2>
-            <h4>{department.description}</h4>
-            <h4>{department.email}</h4>
-            <h4>{department.state}</h4>
-            <h4>{department.telephone}</h4>
-
-            <button onClick={() => editDepartment(department)}>
-              Editar Departamento
-            </button>
-
-            <button onClick={() => editDepartmentAdministrator(department)}>
-              Editar Administrador
-            </button>
-
-            <button onClick={() => deleteDepartmentSelected(department.id)}>
-              Eliminar Departamento
-            </button>
-          </ul>
-        ))}
-      </ul>
-    </div>
-
-*/
